Warnings fixed, added config + setting the best implementation as default, completed other pending tasks, todo.txt fixed
SVN-Revision: 4157

+++ SOC/2007/bigint/boost/bigint/bigint_gmp.hpp
@@ -134,13 +134,13 @@ namespace boost { namespace detail {
 			
 			for (size_t i = 0; i < d_size; ++i)
 			{
-				if (*str < 0 || *str > 127 || digit_value_tab[*str] >= base)
+				if (*str < 0 || *str > 127 || digit_value_tab[static_cast<unsigned int>(*str)] >= base)
 				{
 					d_size = i;
 					break;
 				}
 				
-				d[i] = digit_value_tab[*str++];
+				d[i] = digit_value_tab[static_cast<unsigned int>(*str++)];
 			}
 			
 			if (d_size == 0)
